Added the possibility to type containers inside finely-typed containers
This wasn't in before, since both list and tuple were already taken, so
I didn't see an easy way to add it. Now I realized that it doesn't make
sense to use a list inside another list in the type specifier in the
normal meaning, since the list type specifier needs the previous item to
be a container specifier. In short, list-inside-list didn't have any
meaningful purpose before this, so we can use it to extend typing
deeper. (Don't worry, I probably wont even understand this explanation
myself tomorrow)